requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
boto3==1.38.27
botocore==1.38.27
bs4==0.0.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Deprecated==1.2.18
distro==1.9.0
docstring-parser==0.15
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.5.1
googleapis-common-protos==1.70.0
grpcio==1.71.0
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.3
idna==3.10
importlib_metadata==8.6.1
jiter==0.10.0
jmespath==1.0.1
mcp==1.9.2
numpy==2.2.6
openai==1.82.1
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
packaging==24.2
pinecone==7.0.2
pinecone-plugin-assistant==1.6.1
pinecone-plugin-interface==0.0.7
protobuf==5.29.5
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyMuPDF==1.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
s3transfer==0.13.0
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sse-starlette==2.3.6
starlette==0.46.2
strands-agents==0.1.6
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
wrapt==1.17.2
zipp==3.22.0
